+++ lib/dart_typescript_builder.dart
@@ -16,6 +16,7 @@ export 'src/api_analyzer.dart'
 export 'src/backend/compiler_backend.dart';
 export 'src/backend/dart2js_backend.dart';
 export 'src/backend/dart2wasm_backend.dart';
+export 'src/config.dart';
 export 'src/diagnostics.dart';
 export 'src/dts_generator.dart';
 export 'src/facade_generator.dart';